Refactor `TitleComponent` into multiple components
`Title` is a component for general `Title` setting based on a
`templateString`. `StarTitle` behaves like `Title` but renders the star
icons above the title (and with less customization).
`MemberResultsTitle` inherits from `Title` and manipulates the `state`
to ensure the data needed for the templates exists. `Home` is updated to
use `StarTitle` and `MemberResults` is updated to use
`MemberResultsTitle`. `Results` is changed to more succinctly map from
`legislators` to sections.